+++ src/Components/Nav.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import './Nav.css';
 import { Link } from 'react-router-dom';
 
 const Nav = () => {
