+++ src/Components/Prompt.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import './Prompt.css';
 import PropTypes from 'prop-types';
 
 const Prompt = ({ clickForPrompt, prompt, addToFavorites }) => {
